backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "abap34", email = "abap0002@gmail.com" }
]
dependencies = [
    "fastapi>=0.115.14",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.11.7",
    "sqlalchemy[asyncio]>=2.0.41",
    "alembic>=1.16.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "websockets>=15.0.1",
    "mypy>=1.16.1",
    "pydantic-settings>=2.10.1",
    "aiosqlite>=0.21.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]
